+++ backend/package.json
@@ -13,6 +13,7 @@
     "@types/bcrypt": "^5.0.0",
     "@types/cookie-parser": "^1.4.3",
     "@types/express": "^4.17.17",
+    "@types/jsonwebtoken": "^9.0.1",
     "@types/node": "^18.13.0",
     "ts-node": "^10.9.1",
     "typescript": "^4.9.5"
@@ -25,6 +26,7 @@
     "env-var": "^7.3.0",
     "express": "^4.18.2",
     "http-status-codes": "^2.2.0",
+    "jsonwebtoken": "^9.0.0",
     "zod": "^3.20.6"
   }
 }
