+++ backend/src/index.ts
@@ -4,6 +4,7 @@ import { StatusCodes } from "http-status-codes";
 
 import { backPlay } from "./endpoints/play";
 import { backLogin, backRegister } from "./endpoints/auth";
+import { backPlaylistAdd } from "./endpoints/playlists";
 
 
 function server() {
@@ -24,6 +25,7 @@ function server() {
     backLogin(app);
     backPlay(app);
     backRegister(app);
+    backPlaylistAdd(app);
 }
 
 server()
